Expose handleClearSelection
Add @api decorator to the "Remove selected option" close button. I use lookup with a separate Add button and a datatable. After Adding the selection to the datatable, I need to clear the lookup item. After difficulty finding it in the shadow dom with querySelector and a data-id attribute, this proved to be a much more painless solution! :)